Avoid repeated PDP casting and specify features keyword

+++ tests/unit/test_run_rf_helpers.py
@@ -5,7 +5,7 @@ from farkle.run_rf import plot_partial_dependence
 
 
 def test_plot_partial_dependence(tmp_path):
-    X = pd.DataFrame({"a": range(5), "b": range(5, 10)})
+    X = pd.DataFrame({"a": range(5), "b": range(5, 10)}).astype(float)
     y = pd.Series(range(5))
     model = HistGradientBoostingRegressor(random_state=0)
     model.fit(X, y)
